+++ src/test/java/kitchenpos/bo/OrderBoTest.java
@@ -2,7 +2,7 @@ package kitchenpos.bo;
 
 import kitchenpos.dao.MenuDao;
 import kitchenpos.orders.dao.OrderDao;
-import kitchenpos.dao.OrderLineItemDao;
+import kitchenpos.orders.dao.OrderLineItemDao;
 import kitchenpos.ordertables.dao.OrderTableDao;
 import kitchenpos.orders.model.Order;
 import kitchenpos.orders.model.OrderStatus;
